+++ crates/zizmor/src/audit/artipacked.rs
@@ -1,7 +1,11 @@
 use std::sync::LazyLock;
 
-use github_actions_models::common::{EnvValue, Uses, expr::ExplicitExpr};
+use github_actions_models::common::{
+    EnvValue, Uses,
+    expr::{ExplicitExpr, LoE},
+};
 use itertools::Itertools as _;
+use subfeature::Subfeature;
 
 use super::{Audit, AuditLoadError, audit_meta};
 use crate::{
@@ -105,6 +109,33 @@ impl Artipacked {
             } = &step.body()
             else {
                 continue;
+            };
+
+            let with = match with {
+                LoE::Literal(with) => with,
+                // Emit blanket pedantic finding if the `with:` block cannot be analyzed
+                LoE::Expr(_) => {
+                    findings.push(
+                        Self::finding()
+                            .severity(Severity::Informational)
+                            .confidence(Confidence::High)
+                            .persona(Persona::Pedantic)
+                            .add_location(
+                                step.location()
+                                    .with_keys(["uses".into()])
+                                    .subfeature(Subfeature::new(0, uses.raw()))
+                                    .annotated("this checkout"),
+                            )
+                            .add_location(
+                                step.location()
+                                    .primary()
+                                    .with_keys(["with".into()])
+                                    .annotated("may not set persist-credentials: false"),
+                            )
+                            .build(&step)?,
+                    );
+                    continue;
+                }
             };
 
             if uses.matches("actions/checkout") {