+++ crates/zizmor/tests/integration/e2e/collect.rs
@@ -13,11 +13,11 @@ fn test_fails_incompatible_modes() -> Result<()> {
             .args(["--collect=workflows,actions-only"])
             .input(input_under_test("neutral.yml"))
             .run()?,
-        @"
+        @r"
     🌈 zizmor v@@VERSION@@
     error: `workflows-only` and `actions-only` cannot be combined with other collection modes
 
-    Usage: zizmor [OPTIONS] <INPUTS>...
+    Usage: zizmor [OPTIONS] <INPUT>...
 
     For more information, try '--help'.
     "
@@ -33,7 +33,7 @@ fn test_fails_incompatible_modes() -> Result<()> {
     🌈 zizmor v@@VERSION@@
     error: `workflows-only` and `actions-only` cannot be combined with other collection modes
 
-    Usage: zizmor [OPTIONS] <INPUTS>...
+    Usage: zizmor [OPTIONS] <INPUT>...
 
     For more information, try '--help'.
     "
@@ -49,7 +49,7 @@ fn test_fails_incompatible_modes() -> Result<()> {
     🌈 zizmor v@@VERSION@@
     error: `workflows-only` and `actions-only` cannot be combined with other collection modes
 
-    Usage: zizmor [OPTIONS] <INPUTS>...
+    Usage: zizmor [OPTIONS] <INPUT>...
 
     For more information, try '--help'.
     "